requirements.txt
Django==1.9.7
djangorestframework==3.4.0
django-kronos==0.9
redis==2.10.5
redisco==0.1.4
requests==2.10.0
wheel==0.24.0
